feat: prepare total sums block to inject redux
